+++ src/app/(home)/savedRecipes/page.tsx
@@ -1,7 +1,10 @@
 import { auth } from "@/lib/auth";
 import SavedRecipesView from "@/modules/saved-recipes/views/saved-recipes-view";
+import { api, HydrateClient } from "@/trpc/server";
 import { headers } from "next/headers";
 import { redirect } from "next/navigation";
+
+export const dynamic = "force-dynamic";
 
 export default async function SavedRecipes() {
   const session = await auth.api.getSession({
@@ -12,5 +15,11 @@ export default async function SavedRecipes() {
     redirect("/");
   }
 
-  return <SavedRecipesView />;
+  void api.recipe.getUserRecipes.prefetch();
+
+  return (
+    <HydrateClient>
+      <SavedRecipesView />
+    </HydrateClient>
+  );
 }
